refactor(settings): formalize the settings module boundary
