+++ src/modules/settings/ui/index.ts
@@ -1,1 +1,2 @@
+export { ProfileDetailsForm } from './ProfileDetailsForm'
 export { ThemeSettingsSection } from './ThemeSettingsSection'
