+++ src/pages/index.ts
@@ -5,4 +5,4 @@ export { MyScenesPage } from '@modules/my-scenes'
 export { PlaceholderPage } from './PlaceholderPage'
 export { SceneDetailPage } from '@modules/scene-detail'
 export { ScenesPage } from '@modules/discovery'
-export { SettingsPage } from './SettingsPage'
+export { SettingsPage } from '@modules/settings'
